requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
openai>=1.40.0
google-genai>=1.14.0
python-dotenv>=1.0.1
pydantic>=2.8.2
